Type annotations should not be intermingled with declaration annotations. These changes accommodate that fact.
ATypeElement stands for something that is a type annotation.  Its .type
field is always null.  AElement is used whenever manipulating something
that is not guaranteed to be a type annotation.  This caused changes in the
code to use ATypeElement in many places that AElement was used before.

Update goals to put annotations in correct locations.  Example change:

Fix an indentation problem in the output.
Enhance some debugging output.