Duplicate annotation detection was failing on annotations that had enums as arguments. Prior to my previous patch, it would compare e.g. @Retention(CLASS) (and @java.lang.annotation.Retention(CLASS)) with @Retention(RetentionPolicy.CLASS) and not detect that the annotation already exists.
With the fix, it then tried to compare
@Retention(java.lang.annotation.RetentionPolicy.CLASS) (and
@java.lang.annotation.Retention(java.lang.annotation.RetentionPolicy.CLASS))
with @Retention(RetentionPolicy.CLASS), neither of which match as well.

Rather than generate all possible permutations of package-prefixed and
un-package-prefixed class names (an annotation could declare multiple argument
with enum values), I opt to simply compare only the annotation name, and ignore
the arguments.

It is illegal to duplicate an annotation, so regardless of whether our
arguments match, adding an annotation where one already exists will result in
broken code. A case could be made that a warning should be emitted if we are
instructed to add an annotation for which an existing annotation with different
arguments already exists, however I'm going to limit myself to this partial fix
in the name of expedience.